Refactored "island_configs" dir to "config_templates" dir in blackbox
