+++ envs/monkey_zoo/blackbox/config_templates/base_template.py
@@ -1,4 +1,4 @@
-from envs.monkey_zoo.blackbox.island_configs.config_template import ConfigTemplate
+from envs.monkey_zoo.blackbox.config_templates.config_template import ConfigTemplate
 
 
 # Disables a lot of config values not required for a specific feature test
