+++ envs/monkey_zoo/blackbox/config_templates/drupal.py
