+++ envs/monkey_zoo/blackbox/island_client/island_config_parser.py
@@ -4,7 +4,7 @@ import dpath.util
 from typing_extensions import Type
 
 from envs.monkey_zoo.blackbox.island_client.monkey_island_client import MonkeyIslandClient
-from envs.monkey_zoo.blackbox.island_configs.config_template import ConfigTemplate
+from envs.monkey_zoo.blackbox.config_templates.config_template import ConfigTemplate
 
 
 class IslandConfigParser:
